+++ packages/blockchain-link/src/workers/ripple/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_TIMEOUT = 20 * 1000;
+export const DEFAULT_PING_TIMEOUT = 3 * 60 * 1000;
